hotfix: scoring 필드 Integer → Short (SMALLINT 매핑 수정)
- Flyway V6에서 SMALLINT로 생성한 컬럼과 JPA 타입 불일치 수정
- TrendItem: Integer → Short
- AnalysisService: setter 호출 시 (short) 캐스트
- TrendItemResponse: Map<String, Integer> → Map<String, Short>

+++ mud-backend/src/main/java/com/mud/dto/response/TrendItemResponse.java
@@ -18,7 +18,7 @@ public record TrendItemResponse(
     CategoryResponse category,
     Integer relevanceScore,
     List<String> keywords,
-    Map<String, Integer> scoring,
+    Map<String, Short> scoring,
     String topicTag,
     LocalDateTime publishedAt,
     LocalDateTime crawledAt,
@@ -26,7 +26,7 @@ public record TrendItemResponse(
     String githubLanguage
 ) {
     public static TrendItemResponse from(TrendItem item) {
-        Map<String, Integer> scoring = null;
+        Map<String, Short> scoring = null;
         if (item.getScoringRelevance() != null && item.getScoringTimeliness() != null
                 && item.getScoringActionability() != null && item.getScoringImpact() != null) {
             scoring = Map.of(
